Update with upstream

+++ code/src/java/pcgen/base/formula/DividingFormula.java
@@ -87,7 +87,7 @@ public class DividingFormula implements ReferenceFormula<Integer>
 		 * rounding, and rounding first results in a faster & more accurate
 		 * calculation.
 		 */
-		return Integer.valueOf(numbers[0].intValue() / denominator);
+		return numbers[0].intValue() / denominator;
 	}
 
 	/**
