+++ code/src/java/pcgen/core/term/PCBaseHDTermEvaluator.java
@@ -19,8 +19,6 @@
  * Created 07-May-2014 22:45:18
  *
  * Current Ver: $Revision:$
- * Last Editor: $Author:$
- * Last Edited: $Date:$
  *
  */
 
@@ -33,11 +31,8 @@ import pcgen.core.PlayerCharacter;
  * the character's unmodified number of racial hit dice, as specified by the 
  * MONSTERCLASS tag in the race definition. 
  * 
- * Last Editor: $Author: thpr $
- * Last Edited: $Date: 2012-01-07 18:57:16 +0100 (Sa, 07 Jan 2012) $
  * 
  * @author Stefan Radermacher &lt;zaister@users.sourceforge.net&gt;
- * @version $Revision: 15744 $
  */
 public class PCBaseHDTermEvaluator
 		extends BasePCTermEvaluator implements TermEvaluator {
@@ -58,7 +53,7 @@ public class PCBaseHDTermEvaluator
 	@Override
 	public Float resolve(PlayerCharacter pc)
 	{
-		return (float) pc.getDisplay().getBaseHD();
+		return pc.getDisplay().getBaseHD();
 	}
 
 	/* (non-Javadoc)
