+++ code/src/java/pcgen/gui2/tools/FlippingSplitPane.java
@@ -73,7 +73,6 @@ import pcgen.system.PropertyContext;
  * Locking is also recursive for <code>FlippingSplitPane</code> components.
  *
  * @author &lt;a href="mailto:binkley@alumni.rice.edu"&gt;B. K. Oxley (binkley)&lt;/a&gt;
- * @version $Revision: 1363 $
  */
 public class FlippingSplitPane extends JSplitPane
 {
@@ -193,7 +192,7 @@ public class FlippingSplitPane extends JSplitPane
 	public void setDividerLocation(int location)
 	{
 		PropertyContext context = baseContext.createChildContext(prefsKey);
-		context.setInt(DIVIDER_LOC_PREF_KEY, (Integer) location);
+		context.setInt(DIVIDER_LOC_PREF_KEY, location);
 		if (isLocked())
 		{
 			super.setDividerLocation(getLastDividerLocation());
