+++ code/src/test/pcgen/io/exporttoken/WeaponTokenTest.java
@@ -69,11 +69,8 @@ import pcgen.util.TestHelper;
  * <code>WeaponTokenTest</code> contains tests to verify that the
  * WEAPON token is working correctly.
  *
- * Last Editor: $Author$
- * Last Edited: $Date$
  *
  * @author James Dempsey <jdempsey@users.sourceforge.net>
- * @version $Revision$
  */
 
 public class WeaponTokenTest extends AbstractCharacterTestCase
@@ -314,7 +311,7 @@ public class WeaponTokenTest extends AbstractCharacterTestCase
 		}
 		bite.put(IntegerKey.SLOTS, 0);
 		bite.setQty(Float.valueOf(1));
-		bite.setNumberCarried(Float.valueOf(1));
+		bite.setNumberCarried(1.0f);
 		bite.put(ObjectKey.ATTACKS_PROGRESS, false);
 		bite.getEquipmentHead(1).put(StringKey.DAMAGE, "1d10");
 		bite.getEquipmentHead(1).put(IntegerKey.CRIT_MULT, 2);
@@ -331,7 +328,7 @@ public class WeaponTokenTest extends AbstractCharacterTestCase
 		longbow.addToListFor(ListKey.CAPACITY,
 			new Capacity("Arrow", BigDecimal.ONE));
 		longbow.setQty(Float.valueOf(1));
-		longbow.setNumberCarried(Float.valueOf(1));
+		longbow.setNumberCarried(1.0f);
 		
 		arrow = new Equipment();
 		arrow.setName("Arrow");
